ServerlessRAG.LanceDbAPI/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
fastapi>=0.68.0
lancedb>=0.16.1
pandas
pyarrow
uvicorn
tantivy